Move 'ResultSerializationException' under core.flows
